# contracts/openapi.v1.yaml
openapi: 3.1.0
info:
  title: AI Transaction Authorization Infrastructure API
  version: "1.0.0"
  summary: Deterministic APPROVE/DENY/REVIEW for AI-initiated transactions, before they reach the payment provider.
  description: |
    The decision-plane HTTP surface. A caller POSTs a transaction to `/v1/authorize`
    **before** invoking its payment provider and receives a deterministic, explainable,
    Ed25519-signed decision. **This service never moves, holds, or settles money.**

    ## Two distinct signatures — do not conflate them (ROADMAP A#1)
    - **Request authentication (caller → us): HMAC, symmetric.** Every request carries
      `X-Api-Key`, `X-Signature` (HMAC over a canonical payload), `X-Nonce`, and
      `X-Timestamp`. The caller holds the key; symmetric MAC is appropriate. See the
      `ApiKeyAuth` + `HmacSignature` security schemes below.
    - **Decision signature (us → everyone): Ed25519, asymmetric.** Each `Decision`
      carries a `signature` object produced with our Ed25519 private key. Anyone —
      customer, auditor, regulator — verifies it against the PUBLISHED public key
      (`GET /v1/keys/public`) without holding a secret. HMAC here would let a key-holder
      forge our decisions and is forbidden.

    ## Determinism (ROADMAP A#7)
    A decision is a pure function of `(request, policy_version, evaluated_at, counter_snapshot)`.
    `evaluated_at` and `counter_snapshot` are injected by the orchestrator; predicates never
...
  contact:
...
  license:
...
servers:
...
security:
  - ApiKeyAuth: []
    HmacSignature: []
paths:
  /v1/authorize:
    post:
      tags: [authorize]
      operationId: authorize
      summary: Evaluate a transaction and return a signed decision.
      description: |
        Evaluates the request against the caller's active policy bundle and returns an
        APPROVE / DENY / REVIEW decision with a structured explanation and an Ed25519
        signature. Idempotent per `idempotency_key`: replaying an identical request with
        the same key returns the same decision and creates no additional side effect.
      parameters:
        - $ref: "#/components/parameters/ApiKey"
        - $ref: "#/components/parameters/Signature"
        - $ref: "#/components/parameters/Nonce"
        - $ref: "#/components/parameters/Timestamp"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./schemas/authorize-request.schema.json"
            examples:
              vendorPayment:
                summary: Procurement agent paying a vendor
                externalValue: ./examples/authorize-request.example.json
      responses:
        "200":
          description: A decision was produced (APPROVE, DENY, or REVIEW — all are 200).
          headers:
            X-Decision-Id:
              description: The decision_id, mirrored for logging/correlation.
              schema: { type: string }
            X-Shadow:
              description: |
                Present and "true" when the decision is advisory (a shadow-mode key):
                the caller must NOT enforce it. Mirrors the signed `shadow` field.
              schema: { type: string, enum: ["true"] }
          content:
            application/json:
              schema:
                $ref: "./schemas/decision.schema.json"
              examples:
                approve:
                  summary: APPROVE
                  externalValue: ./examples/decision-approve.example.json
                deny:
                  summary: DENY (monthly budget exceeded)
                  externalValue: ./examples/decision-deny.example.json
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/ReplayDetected"
        "422":
          $ref: "#/components/responses/Unprocessable"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/Unavailable"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: |
        Per-key API credential, prefix `azn_live_` / `azn_test_`. Only `sha256(key)` is
        stored server-side; comparison is constant-time (TRD §11).
    HmacSignature:
      type: apiKey
      in: header
      name: X-Signature
      description: |
        HMAC (symmetric) over the canonical request payload using the API key secret, paired
        with `X-Nonce` and `X-Timestamp`. Verified constant-time; timestamp skew beyond the
        configured window and reused nonces are rejected. This authenticates the CALLER and is
        entirely separate from the Ed25519 signature on the returned Decision.
  parameters:
    ApiKey:
      name: X-Api-Key
      in: header
      required: true
      schema: { type: string, pattern: "^azn_(live|test)_[A-Za-z0-9_-]+$" }
    Signature:
      name: X-Signature
      in: header
      required: true
      description: HMAC-SHA256 of the canonical payload, hex or base64url.
      schema: { type: string }
    Nonce:
      name: X-Nonce
      in: header
      required: true
      description: Unique per request; cached per key to reject replays.
      schema: { type: string, minLength: 8, maxLength: 128 }
    Timestamp:
      name: X-Timestamp
      in: header
      required: true
      description: RFC 3339 request time; rejected if skew exceeds the configured window.
      schema: { type: string, format: date-time }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        Optional transport mirror of the body `idempotency_key`. When both are present they
        must match. Makes authorize retry-safe.
      schema: { type: string, minLength: 8, maxLength: 128 }
...
  responses:
    ValidationError:
      description: Malformed request (schema/shape).
      content:
        application/problem+json:
          schema: { $ref: "./schemas/error.schema.json" }
    Unauthorized:
      description: Missing/invalid API key or bad HMAC signature.
      content:
        application/problem+json:
          schema: { $ref: "./schemas/error.schema.json" }
    Forbidden:
      description: Key revoked or lacks scope.
      content:
        application/problem+json:
          schema: { $ref: "./schemas/error.schema.json" }
...
    ReplayDetected:
      description: Nonce already seen for this key.
      content:
        application/problem+json:
          schema: { $ref: "./schemas/error.schema.json" }
    Unprocessable:
      description: Well-formed but semantically invalid (e.g. currency/amount inconsistency).
      content:
        application/problem+json:
          schema: { $ref: "./schemas/error.schema.json" }
    RateLimited:
      description: Per-key rate or burst limit exceeded.
      headers:
        Retry-After:
          description: Seconds to wait before retrying.
          schema: { type: integer }
      content:
        application/problem+json:
          schema: { $ref: "./schemas/error.schema.json" }
    Unavailable:
      description: Decision plane could not serve (e.g. no last-known-good bundle). Never a silent guess.
      content:
        application/problem+json:
          schema: { $ref: "./schemas/error.schema.json" }
...